Fix text rendering issue in webkit/blink under certain widths (#13)
In webkit and blink, rendering the cursor alongside the text as it
animates sometimes causes the text to stop rendering when it reaches
a new line break, even though the underlying DOM /does/ contain
the text. This seems to happen when the space available for the text is
at a specific width that makes it so the line break happens within a
word.

Using dev tools, when in this state, if you modify the dom or any style,
it immediately renders all of the text in its correct position.
Given that observation, this is a hackish solutions that re-renders the
cursor every time a character is rendered, just to ensure that the text
is rendered correctly.

See issue #13 and for more details