Update README.md
The link to the css file is broken and i found the scss one. Struggled at first a little to use it but then was much easier then implementing my own blink since it comes out of the box. I wanted to use css and simply converted it from scss to css online. And editing it it's simple too, wanted to make it look more like the one from ubuntu terminal.